Vectoryn/search/requirements.txt
fastapi>=0.100.0
uvicorn[standard]>=0.23.0
pydantic>=2.0.0
qdrant-client==1.8.0
torch==2.2.2
transformers==4.40.0
python-dotenv
groq